pyproject.toml
[project]
authors = [
  { name = "Denrei Keith De Jesus", email = "denreikeith@gmail.com" },
  { name = "Stephanie Palero", email = "s.palero.526878@umindanao.edu.ph" },
  { name = "Joshua Roi Nomus", email = "j.nomus.523625@umindanao.edu.ph" },
]
name = "MiRAG"
version = "0.1.0"
description = "Mind the (Context) Gap: Reducing AI Hallucination by Bridging Long-Context Retrieval and Information Refinement through Mindful RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.2.0",
    "google-genai>=1.0.0",
    "huggingface-hub>=0.28.1",
    "llama-index>=0.12.19",
    "llama-index-embeddings-huggingface>=0.5.1",
    "llama-index-llms-gemini>=0.4.10",
    "llama-index-tools-tavily-research>=0.3.0",
    "llama-index-utils-workflow>=0.3.0",
    "nest-asyncio>=1.6.0",
    "numpy>=2.2.2",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.48.3",
    "uvloop>=0.21.0",
    "torch",
    # "pytorch-triton-rocm",
    "chromadb>=0.6.3",
    "llama-index-vector-stores-chroma>=0.4.1",
    "loguru>=0.7.3",
    "uvicorn>=0.34.0",
    "fastapi>=0.115.11",
    "llama-index-readers-huggingface-fs>=0.3.1",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-embeddings-langchain>=0.3.0",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "boto3>=1.37.13",
    "authlib>=1.5.1",
    "jwt>=1.3.1",
    "itsdangerous>=2.2.0",
    "pynamodb>=6.0.2",
    "python-jose>=3.4.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.10.6",
    "llama-index-vector-stores-pinecone>=0.4.5",
    "llama-index-vector-stores-dynamodb>=0.3.0",
    "llama-index-storage-docstore-dynamodb>=0.3.0",
    "llama-index-storage-index-store-dynamodb-store>=0.4.0",
    "llama-index-storage-docstore-redis>=0.3.0",
    "llama-index-storage-index-store-redis>=0.4.0",
    "llama-index-llms-azure-openai>=0.3.2",
    "llama-index-llms-deepseek>=0.1.1",
    "ragas>=0.2.14",
    "llama-index-llms-ollama>=0.5.4",
    "rouge-score>=0.1.2",
    "jupyter>=1.1.1",
    "mlx>=0.25.1",
    "mlx-lm>=0.24.0",
    "python-multipart>=0.0.20",
    "s3fs>=0.4.2",
    "firebase-admin>=6.9.0",
]

[dependency-groups]
lint =  [
"ruff>=0.9.5"
]
dev = [
    "icecream>=2.1.4",
    "peft>=0.14.0",
    "pre-commit>=4.1.0",
]

[project.scripts]
mirag = "mirag.main:run"
api = "api.main:run"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


# [tool.uv.sources]
# torch = [
#   { index = "pytorch-rocm" },
# ]
# pytorch-triton-rocm = [
#   { index = "pytorch-rocm" },
# ]

# [[tool.uv.index]]
# name = "pytorch-rocm"
# url = "https://download.pytorch.org/whl/rocm6.2"
# explicit = true


[tool.ruff]
line-length = 120
target-version = "py312"
